@@ -1,20 +1,20 @@
 [tool.poetry]
 name = "vidypy"
-version = "0.1.0"
+version = "0.2.0"
 description = ""
 authors = ["Iván Salazar <ivangio.salazar@gmail.com>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.11"
-python-telegram-bot = "^20.3"
+python = "^3.12"
+python-telegram-bot = "^21.1.1"
 yt-dlp = "^2024.3.10"
-validators = "^0.22.0"
+validators = "^0.28.1"
 
 
 [tool.poetry.group.dev.dependencies]
 autoimport = "^1.3.3"
-ruff = "^0.1.6"
+ruff = "^0.4.4"
 
 [build-system]
 requires = ["poetry-core"]
